feat: case-insensitive package aliases
Packages can now have multiple aliases which redirect to the package's
canonical name.

If a package has a `lake-manifest.json`, the package's canonical name is
the `name` listed in the manifest. Otherwise, Reservoir falls back to
the name of the GitHub repository it is sourced from, Aliases are
represented in the by a file (rather than a directory) containing the
canonical path to the package.

The path of the package in the index is now case-insensitive. Only the
metadata of a package preserves the casing of a package's full name.
This is done to make the index compatible with case-insensitive file
systems. This means the index can no longer contain multiple packages
with names that only differ in casing.

During an update (or CI), Reservoir will refresh all previously indexed
packages and move them to their proper canonical path, creating aliases
along the way. If there is already a package at the new name (e.g.,
because they were both indexed before this change), Reservoir will merge
the packages into one (with the canonically named package taking
precedence).

CI Keyword(s): UPDATE-INDEX