feat: toolchain metadata & related cleanup
Site toolchains now carry additional metadata. This metadata is used to
accomplish a few things:

* Sort toolchains by the date the release was created.
* The main page now displays the latest stable Lean toolchain, rather
than simply the latest one. Clicking on the toolchain will take the user
to the toolchain's release notes.
* A package is considered to have an up-to-date build (a green rather
than a gold checkmark) if the package toolchain with the same or newer
minor version as the latest stable release. For example, if the latest
stable is `4.6.1`, a package which builds on any of `4.6.0-rc1`,
`4.6.0`, `4.6.1`, or `v4.7.0-rc1` is considered up-to-date.

This also includes some general cleanup / refactoring of related code. 

Closes #12.