fix: `package.json` per netlify function
Testing indicates that this fixes the recently appearing function upload
error in deploys. It is not an ideal solution. It makes importing
utilities from the main codebase more finicky and, if Reservoir had more
functions, requires a `package.json` for each. However, this is a
critical blocker, so a unideal solution is fine for now.