feat: build API endpoints + netlify API refactor
Adds new API endpoints for retrieving package builds, versions, and
barrels (S3 cloud build archives. In service of this, It converts the
previous single Netlify API function into a full API with a (H3) router.
With this reorganization, the API and website are now separate NPM
workspaces within the larger monorepo. This PR also adds a
`netlify.toml` to simplify deployment configuration.